Resolver module separation
