+++ crates/cairo-lang-semantic/src/items/extern_type.rs
@@ -7,7 +7,7 @@ use super::generics::semantic_generic_params;
 use crate::db::SemanticGroup;
 use crate::diagnostic::SemanticDiagnosticKind::*;
 use crate::diagnostic::SemanticDiagnostics;
-use crate::resolve_path::Resolver;
+use crate::resolve::Resolver;
 use crate::substitution::SemanticRewriter;
 use crate::{GenericParam, SemanticDiagnostic};
