+++ crates/cairo-lang-semantic/src/items/structure.rs
@@ -13,7 +13,7 @@ use super::generics::semantic_generic_params;
 use crate::db::SemanticGroup;
 use crate::diagnostic::SemanticDiagnosticKind::*;
 use crate::diagnostic::SemanticDiagnostics;
-use crate::resolve_path::{ResolvedLookback, Resolver};
+use crate::resolve::{ResolvedItems, Resolver};
 use crate::substitution::{GenericSubstitution, SemanticRewriter, SubstitutionRewriter};
 use crate::types::{resolve_type, ConcreteStructId};
 use crate::{semantic, SemanticDiagnostic};
@@ -31,7 +31,7 @@ pub struct StructDeclarationData {
     diagnostics: Diagnostics<SemanticDiagnostic>,
     generic_params: Vec<semantic::GenericParam>,
     attributes: Vec<Attribute>,
-    resolved_lookback: Arc<ResolvedLookback>,
+    resolved_lookback: Arc<ResolvedItems>,
 }
 
 /// Query implementation of [crate::db::SemanticGroup::priv_struct_declaration_data].
@@ -60,7 +60,7 @@ pub fn priv_struct_declaration_data(
     )?;
 
     let attributes = ast_attributes_to_semantic(syntax_db, struct_ast.attributes(syntax_db));
-    let resolved_lookback = Arc::new(resolver.lookback);
+    let resolved_lookback = Arc::new(resolver.resolved_items);
 
     // Check fully resolved.
     if let Some((stable_ptr, inference_err)) = resolver.inference.finalize() {
@@ -104,7 +104,7 @@ pub fn struct_attributes(db: &dyn SemanticGroup, struct_id: StructId) -> Maybe<V
 pub fn struct_declaration_resolved_lookback(
     db: &dyn SemanticGroup,
     struct_id: StructId,
-) -> Maybe<Arc<ResolvedLookback>> {
+) -> Maybe<Arc<ResolvedItems>> {
     Ok(db.priv_struct_declaration_data(struct_id)?.resolved_lookback)
 }
 
@@ -114,7 +114,7 @@ pub fn struct_declaration_resolved_lookback(
 pub struct StructDefinitionData {
     diagnostics: Diagnostics<SemanticDiagnostic>,
     members: OrderedHashMap<SmolStr, Member>,
-    resolved_lookback: Arc<ResolvedLookback>,
+    resolved_lookback: Arc<ResolvedItems>,
 }
 #[derive(Clone, Debug, Hash, PartialEq, Eq, DebugWithDb, SemanticObject)]
 #[debug_db(dyn SemanticGroup + 'static)]
@@ -161,7 +161,7 @@ pub fn priv_struct_definition_data(
         }
     }
 
-    let resolved_lookback = Arc::new(resolver.lookback);
+    let resolved_lookback = Arc::new(resolver.resolved_items);
 
     // Check fully resolved.
     if let Some((stable_ptr, inference_err)) = resolver.inference.finalize() {
@@ -197,7 +197,7 @@ pub fn struct_members(
 pub fn struct_definition_resolved_lookback(
     db: &dyn SemanticGroup,
     struct_id: StructId,
-) -> Maybe<Arc<ResolvedLookback>> {
+) -> Maybe<Arc<ResolvedItems>> {
     Ok(db.priv_struct_declaration_data(struct_id)?.resolved_lookback)
 }
