+++ crates/cairo-lang-semantic/src/items/us.rs
@@ -8,7 +8,7 @@ use cairo_lang_utils::Upcast;
 use crate::db::SemanticGroup;
 use crate::diagnostic::SemanticDiagnosticKind::*;
 use crate::diagnostic::{NotFoundItemType, SemanticDiagnostics};
-use crate::resolve_path::{ResolvedGenericItem, ResolvedLookback, Resolver};
+use crate::resolve::{ResolvedGenericItem, ResolvedItems, Resolver};
 use crate::SemanticDiagnostic;
 
 #[derive(Clone, Debug, PartialEq, Eq, DebugWithDb)]
@@ -16,7 +16,7 @@ use crate::SemanticDiagnostic;
 pub struct UseData {
     diagnostics: Diagnostics<SemanticDiagnostic>,
     resolved_item: Maybe<ResolvedGenericItem>,
-    resolved_lookback: Arc<ResolvedLookback>,
+    resolved_lookback: Arc<ResolvedItems>,
 }
 
 /// Query implementation of [crate::db::SemanticGroup::priv_use_semantic_data].
@@ -36,7 +36,7 @@ pub fn priv_use_semantic_data(db: &(dyn SemanticGroup), use_id: UseId) -> Maybe<
         &use_ast.name(syntax_db),
         NotFoundItemType::Identifier,
     );
-    let resolved_lookback = Arc::new(resolver.lookback);
+    let resolved_lookback = Arc::new(resolver.resolved_items);
 
     // Check fully resolved.
     if let Some((stable_ptr, inference_err)) = resolver.inference.finalize() {
@@ -68,7 +68,7 @@ pub fn priv_use_semantic_data_cycle(
     Ok(UseData {
         diagnostics: diagnostics.build(),
         resolved_item: err,
-        resolved_lookback: Arc::new(ResolvedLookback::default()),
+        resolved_lookback: Arc::new(ResolvedItems::default()),
     })
 }
 
@@ -81,10 +81,7 @@ pub fn use_semantic_diagnostics(
 }
 
 /// Query implementation of [crate::db::SemanticGroup::use_resolved_lookback].
-pub fn use_resolved_lookback(
-    db: &dyn SemanticGroup,
-    use_id: UseId,
-) -> Maybe<Arc<ResolvedLookback>> {
+pub fn use_resolved_lookback(db: &dyn SemanticGroup, use_id: UseId) -> Maybe<Arc<ResolvedItems>> {
     Ok(db.priv_use_semantic_data(use_id)?.resolved_lookback)
 }
